Added an option for higher-power WiFi (noms up battery!)

+++ ESP32/MQTT_SubHandler/include/xasin/mqtt/Handler.h
@@ -48,7 +48,7 @@ protected:
 	std::string status_msg;
 
 public:
-	static void start_wifi(const char *SSID, const char *PSWD, uint8_t psMode = 0);
+	static void start_wifi(const char *SSID, const char *PSWD, int psMode = 0);
 	static void start_wifi_enterprise(const char *SSID, const char *domain, const char *identity, const char *anonymousIdentity, const char *password);
 	static void try_wifi_reconnect(system_event_t *event);
